[Core] Move resourcetracker folder into resources, and move RegenResourceCapTracker into resourcetracker
