+++ src/parser/druid/balance/modules/resourcetracker/AstralPowerTracker.js
@@ -1,5 +1,5 @@
 import RESOURCE_TYPES from 'game/RESOURCE_TYPES';
-import ResourceTracker from 'parser/shared/modules/resourcetracker/ResourceTracker';
+import ResourceTracker from 'parser/shared/modules/resources/resourcetracker/ResourceTracker';
 import SPELLS from 'common/SPELLS';
 
 const WARRIOR_OF_ELUNE_MULTIPLIER = 0.4;
