+++ src/parser/druid/feral/modules/features/EnergyTracker.js
@@ -1,6 +1,6 @@
 import SPELLS from 'common/SPELLS';
 import RESOURCE_TYPES from 'game/RESOURCE_TYPES';
-import ResourceTracker from 'parser/shared/modules/resourcetracker/ResourceTracker';
+import ResourceTracker from 'parser/shared/modules/resources/resourcetracker/ResourceTracker';
 
 class EnergyTracker extends ResourceTracker {
   constructor(...args) {
