+++ src/parser/hunter/marksmanship/modules/resources/Focus.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import Analyzer from 'parser/core/Analyzer';
 import SPELLS from 'common/SPELLS';
 import SpellLink from 'common/SpellLink';
-import resourceSuggest from 'parser/shared/modules/resourcetracker/ResourceSuggest';
+import resourceSuggest from 'parser/shared/modules/resources/resourcetracker/ResourceSuggest';
 import FocusTracker from 'parser/hunter/shared/modules/resources/FocusTracker';
 
 class Focus extends Analyzer {
