+++ src/parser/hunter/shared/modules/resources/FocusTracker.tsx
@@ -1,7 +1,7 @@
 import RESOURCE_TYPES from 'game/RESOURCE_TYPES';
-import ResourceTracker from 'parser/shared/modules/resourcetracker/ResourceTracker';
+import ResourceTracker from 'parser/shared/modules/resources/resourcetracker/ResourceTracker';
 import SPELLS from 'common/SPELLS';
-import { EnergizeEvent } from 'parser/core/Events';
+import { CastEvent, EnergizeEvent } from 'parser/core/Events';
 import { CHIM_REGEN } from 'parser/hunter/shared/constants';
 import { AOTW_REGEN, BARBED_SHOT_FOCUS_REGEN_BUFFS, BARBED_SHOT_REGEN, BEAST_MASTERY_SPELLS_WITHOUT_WASTE } from 'parser/hunter/beastmastery/constants';
 
@@ -34,6 +34,11 @@ class FocusTracker extends ResourceTracker {
     }
 
     this._applyBuilder(spellId, this.getResource(event), gain, waste, event.timestamp);
+  }
+
+  //We're not interested in the fabricated events that are created prepull, as these tend to add focus costs things such as potion usage, Bestial Wrath, Trueshot or the likes whereas these don't cost focus.
+  shouldProcessCastEvent(event: CastEvent) {
+    return !event.__fabricated && !!this.getResource(event);
   }
 }
