+++ src/parser/monk/mistweaver/modules/resourcetracker/SoothingMistTracker.js
@@ -1,5 +1,5 @@
 import RESOURCE_TYPES from 'game/RESOURCE_TYPES';
-import ResourceTracker from 'parser/shared/modules/resourcetracker/ResourceTracker';
+import ResourceTracker from 'parser/shared/modules/resources/resourcetracker/ResourceTracker';
 
 class SoothingMistTracker extends ResourceTracker {
   constructor(...args) {
