+++ src/parser/rogue/assassination/modules/core/ComboPoints.js
@@ -3,7 +3,7 @@ import React from 'react';
 import Analyzer from 'parser/core/Analyzer';
 import SPELLS from 'common/SPELLS';
 import SpellLink from 'common/SpellLink';
-import resourceSuggest from 'parser/shared/modules/resourcetracker/ResourceSuggest';
+import resourceSuggest from 'parser/shared/modules/resources/resourcetracker/ResourceSuggest';
 
 import ComboPointTracker from '../../../shared/resources/ComboPointTracker';
