+++ src/parser/rogue/subtlety/modules/core/ComboPoints.js
@@ -3,7 +3,7 @@ import React from 'react';
 import Analyzer from 'parser/core/Analyzer';
 import SPELLS from 'common/SPELLS/index';
 import SpellLink from 'common/SpellLink';
-import resourceSuggest from 'parser/shared/modules/resourcetracker/ResourceSuggest';
+import resourceSuggest from 'parser/shared/modules/resources/resourcetracker/ResourceSuggest';
 
 import ComboPointTracker from '../../../shared/resources/ComboPointTracker';
 
@@ -15,7 +15,7 @@ class ComboPoints extends Analyzer {
   makeExtraSuggestion(spell) {
     return <>Avoid wasting combo points when casting <SpellLink id={spell.id} /> </>;
   }
-  
+
   get comboPointThresholds() {
     return {
       actual: this.comboPointTracker.wasted / this.comboPointTracker.generated,
