+++ src/parser/shared/modules/resources/resourcetracker/RegenResourceCapTracker.js
