+++ src/parser/shared/modules/resources/resourcetracker/ResourceSuggest.js
