+++ src/parser/warlock/affliction/modules/soulshards/SoulShardTracker.js
@@ -1,4 +1,4 @@
-import ResourceTracker from 'parser/shared/modules/resourcetracker/ResourceTracker';
+import ResourceTracker from 'parser/shared/modules/resources/resourcetracker/ResourceTracker';
 import RESOURCE_TYPES from 'game/RESOURCE_TYPES';
 
 class SoulShardTracker extends ResourceTracker {
