+++ src/parser/warlock/destruction/modules/soulshards/SoulShardTracker.js
@@ -1,4 +1,4 @@
-import ResourceTracker from 'parser/shared/modules/resourcetracker/ResourceTracker';
+import ResourceTracker from 'parser/shared/modules/resources/resourcetracker/ResourceTracker';
 import HIT_TYPES from 'game/HIT_TYPES';
 import Enemies from 'parser/shared/modules/Enemies';
 import EventFilter from 'parser/core/EventFilter';
